Fix sections (#55)
* Set result context to proper value

When section() is used as a context manager, we need to save the proper
ResultContext so the test report points to the line where the section is used.
Previously, we pointed to contextlib, which was a mistake.

* Fix section results entries to point to proper script lines

When the section was used as a function decorator, the corresponding
log entry incorrectly pointed to the contextlib library instead of the
actual location where the function was called. This has been fixed.
Now, both use cases—decorator and context manager—point to the correct
script lines.